Router init and create separate main search component
