+++ frontend/steam-comments/package.json
@@ -10,6 +10,7 @@
     "@vue/composition-api": "^1.4.2",
     "core-js": "^3.6.5",
     "vue": "^2.6.11",
+    "vue-router": "^3.5.3",
     "vuetify": "^2.4.0"
   },
   "devDependencies": {
