+++ frontend/steam-comments/src/main.ts
@@ -1,4 +1,5 @@
 import "./plugins/compositionApi";
+import router from "@/router/router";
 import Vue from "vue";
 import App from "./App.vue";
 import vuetify from "./plugins/vuetify";
@@ -8,4 +9,5 @@ Vue.config.productionTip = false;
 new Vue({
   vuetify,
   render: (h) => h(App),
+  router,
 }).$mount("#app");
